+++ lib/ui/screens/supplier_products_screen.dart
@@ -1229,7 +1229,7 @@ class SupplierProductsScreenWidget extends StatelessWidget {
                                   mainAxisAlignment: MainAxisAlignment.start,
                                   crossAxisAlignment: CrossAxisAlignment.start,
                                   children: [
-                                    int.parse(productStock) > 0  && lowStock.isEmpty ? 0.width : productStock == '0' && lowStock.isNotEmpty ? Text(
+                                    double.parse(productStock) > 0  && lowStock.isEmpty ? 0.width : productStock == '0' && lowStock.isNotEmpty ? Text(
                                       AppLocalizations.of(context)!
                                           .out_of_stock1,
                                       style: AppStyles.rkBoldTextStyle(
